Major change.
All the algorithms were packed in one model_handles file. For this effect a new technique was learned and implemented, which allowed for many functions and subroutines to be present all-in-one .m file. The algorithms remain exactly the same. 

The GUI was completely restructured. Its files are stored in a GUI folder. The GUI is incomplete.

The input and output data was better organized in folders.